% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCEB.moran.plot.R
\name{HCEB.moran.plot}
\alias{HCEB.moran.plot}
\title{Moran scatterplot with Heteroscedasticity-Consistent Empirical Bayes rate}
\usage{
HCEB.moran.plot(n, x, se, listw, zero.policy = NULL, spChk = NULL,
  labels = NULL, xlab = NULL, ylab = NULL, ...)
}
\arguments{
\item{n}{a numeric vector of counts of cases (numerator)}

\item{x}{a numeric vector of populations at risk (denominator)}

\item{se}{a numeric vector of sampling standard errors of corresponding populations at risk}

\item{listw}{a listw object created for example by nb2listw}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}

\item{spChk}{should the data vector names be checked against the spatial objects for identity integrity, TRUE, or FALSE, default NULL to use get.spChkOption()}

\item{labels}{character labels for points with high influence measures, if set to FALSE, no labels are plotted for points with large influence}
}
\description{
A plot of HC-EB spatial areal rate data against its spatially lagged values.
}
\references{
Jung, PH, Thill J-C, Issel M 2018 Spatial Autocorrelation Statistics of Areal Prevalence Rates under High Uncertainty in Denominator Data, Geographical Analysis
}
\keyword{HCEB}
