% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCEB.est.R
\name{HCEB.est}
\alias{HCEB.est}
\title{Heteroscedastic Consistent Empirical Bayes rate}
\usage{
HCEB.est(x, n, k, b = 0)
}
\arguments{
\item{x}{a numeric vector of counts of cases}

\item{n}{a numeric vector of populations at risk}

\item{k}{a numeric vector of margin of errors of x}

\item{b}{default 0, base populations-at-risk}
}
\description{
Compute HC-EB rate estimates for rates smoothed to the overall mean by their denominator data uncertainty.
}
\references{
#####, ##### and #####. 2018. Small area adjustment of spatial autocorrelation estimators for heteroscedastic sampling errors, Geographical Analysis (under review)
}
\keyword{HCEB}
