% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCEB.est.R
\name{HCEB.est}
\alias{HCEB.est}
\title{Heteroscedastic Consistent Empirical Bayes rate}
\usage{
HCEB.est(n, x, se)
}
\arguments{
\item{n}{a numeric vector of counts of cases (numerator)}

\item{x}{a numeric vector of populations at risk (denominator)}

\item{se}{a numeric vector of sampling standard errors of corresponding populations at risk}
}
\description{
Compute HC-EB rate estimates for rates smoothed to the overall mean by their denominator data uncertainty.
}
\references{
Jung, PH, Thill J-C, Issel M 2018 Spatial Autocorrelation Statistics of Areal Prevalence Rates under High Uncertainty in Denominator Data, Geographical Analysis
}
\keyword{HCEB}
