% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCEB.moran.mc.R
\name{HCEB.moran.mc}
\alias{HCEB.moran.mc}
\title{Permutation test for Global Heteroscedastic Consistent Empirical Bayes Moran's I}
\usage{
HCEB.moran.mc(x, n, k, b = 0, listw, sig = 0.05, nsim,
  zero.policy = NULL)
}
\arguments{
\item{x}{a numeric vector of counts of cases}

\item{n}{a numeric vector of populations at risk}

\item{k}{a numeric vector of margin of errors of x}

\item{b}{default 0, base populations-at-risk}

\item{listw}{a listw object created for example by nb2listw}

\item{sig}{default 0.05, pseudo significance level for permutation test}

\item{nsim}{number of permutations}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}
}
\description{
A permutation test for Global HC-EB Moran's I statistics by using nsim random permutations of HC-EB rate.
}
\references{
#####, ##### and #####. 2018. Small area adjustment of spatial autocorrelation estimators for heteroscedastic sampling errors, Geographical Analysis (under review)
}
\keyword{HCEB}
