% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HCEB.moran.R
\name{HCEB.moran}
\alias{HCEB.moran}
\title{Global Heteroscedastic Consistent Empirical Bayes Moran's I estimator}
\usage{
HCEB.moran(n, x, se, listw, zero.policy = NULL)
}
\arguments{
\item{n}{a numeric vector of counts of cases (numerator)}

\item{x}{a numeric vector of populations at risk (denominator)}

\item{se}{a numeric vector of sampling standard errors of corresponding populations at risk}

\item{listw}{a listw object created for example by nb2listw}

\item{zero.policy}{default NULL, use global option value; if TRUE assign zero to the lagged value of zones without neighbours, if FALSE assign NA}
}
\description{
Compute Global HC-EB Moran's I. Global Moran's I of areal prevalence rates with uncertainty in denominator.
}
\references{
Jung, PH, Thill J-C, Issel M 2018 Spatial Autocorrelation Statistics of Areal Prevalence Rates under High Uncertainty in Denominator Data, Geographical Analysis
}
\keyword{HCEB}
